Printing out text with bs4
We are now able to get the page's html, as well as make it more visible to understand what is being looked at. Its possible a framework is being used to be able to generate all the tables shown on the site. More research needs to be done in order to figure out how to properly get all the tables that are generated, as well as be able to scrape the problems that are being shown.